Move things to right folder + solve scrolling bug.
